move InterfaceCollection to utils
