+++ src/NetMQ.Tests/NetMQ3.5.Tests.csproj
@@ -81,7 +81,6 @@
     <Compile Include="ReqRepTests.cs" />
     <Compile Include="SocketTests.cs" />
     <Compile Include="StreamTests.cs" />
-    <Compile Include="zmq\YQueueTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\NetMQ\NetMQ3.5.csproj">
